+++ src/cli/actions/clear.rs
@@ -1,4 +1,4 @@
-use indexmap::IndexSet;
+use indexmap::{IndexMap, map::Entry};
 
 use crate::{
     cli::{
@@ -7,22 +7,29 @@ use crate::{
         prompt::{Prompt, PromptAnswer, PromptFlags},
     },
     fs::{abs::AbsPathStr, rel::RelPathStr},
-    prof::{ProfileKind, module::ModuleEntry, runner::RunnerEntry},
+    prof::{ProfileKind, module::ModulePolicy},
 };
 
 fn resolve<'a>(
-    relpaths: impl Iterator<Item = &'a RelPathStr>,
+    relpaths: impl Iterator<Item = (&'a RelPathStr, bool)>,
     dir: &AbsPathStr,
-    entries: &mut IndexSet<AbsPathStr>,
+    entries: &mut IndexMap<AbsPathStr, bool>,
 ) -> anyhow::Result<()> {
-    for path in relpaths {
+    for (path, ignored) in relpaths {
         for p in path.to_abs(dir)?.all_files_ord()? {
             let p = if p.is_file() {
                 p.canonicalize()?
             } else {
                 continue;
             };
-            entries.insert(p);
+            match entries.entry(p) {
+                Entry::Occupied(mut e) => {
+                    e.insert(*e.get() || ignored);
+                }
+                Entry::Vacant(e) => {
+                    e.insert(ignored);
+                }
+            };
         }
     }
     Ok(())
@@ -35,7 +42,7 @@ impl Cli {
                 let run_dir = &ctx.paths[&Paths::Run];
                 let backup_dir = &ctx.paths[&Paths::Backup];
                 let root_dir = &ctx.paths[&Paths::Root];
-                let mut entries = IndexSet::new();
+                let mut entries = IndexMap::new();
                 let prompt = Prompt::new(
                     PromptAnswer::all() & !PromptAnswer::DIFF,
                     PromptFlags::new(self.assume_no, self.assume_yes, self.list),
@@ -47,12 +54,16 @@ impl Cli {
                     match ctx.item.kind() {
                         ProfileKind::Module(module) => {
                             let this_backup_dir = backup_dir.join(ctx.item.id_or(ctx.name))?;
-                            let relpaths = module.entries().iter().map(ModuleEntry::path);
+                            let relpaths = module
+                                .entries()
+                                .iter()
+                                .map(|e| (e.path(), e.policy() == &ModulePolicy::Ignore));
                             resolve(relpaths, &this_backup_dir, &mut entries)?;
                         }
                         ProfileKind::Runner(runner) => {
                             let this_runner_dir = run_dir.join(ctx.item.id_or(ctx.name))?;
-                            let relpaths = runner.entries().iter().map(RunnerEntry::path);
+                            let relpaths = runner.entries().iter().map(|e| (e.path(), false));
+
                             resolve(relpaths, &this_runner_dir, &mut entries)?;
                         }
                         _ => {}
@@ -68,7 +79,17 @@ impl Cli {
                         } else {
                             continue;
                         };
-                        if !entries.contains(&file) {
+                        if let Some(ignored) = entries.get(&file) {
+                            if *ignored {
+                                let relpath = file.to_rel(root_dir)?;
+                                ctx.col.output_path(&relpath, ctx.col.output_path);
+                                prompt.handled_prompt_available(
+                                    "Do you really want to delete ignored file?",
+                                    &[&file],
+                                    || file.purge_path(),
+                                )?;
+                            }
+                        } else {
                             let relpath = file.to_rel(root_dir)?;
                             ctx.col.output_path(&relpath, ctx.col.output_path);
                             prompt.handled_prompt_available(